requirements.txt
fastapi
uvicorn
pyperclip
python-dotenv
pydantic
jinja2
starlette
groq
instructor